s3 업다운 리팩토링
