--- dokky/src/main/java/org/my/s3/S3Util.java
@@ -1,150 +1,0 @@
-/*package org.my.s3;
-import java.io.ByteArrayInputStream;
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.util.List;
-
-import com.amazonaws.ClientConfiguration;
-import com.amazonaws.Protocol;
-import com.amazonaws.auth.AWSCredentials;
-import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3Client;
-import com.amazonaws.services.s3.model.Bucket;
-import com.amazonaws.services.s3.model.GeneratePresignedUrlRequest;
-import com.amazonaws.services.s3.model.ObjectMetadata;
-
-public class S3Util {
-
-	private String accessKey = "엑세스키"; // 엑세스 키
-	private String secretKey = "보안 엑세스키"; // 보안 엑세스 키
-
-	private AmazonS3 conn;
-
-	public S3Util() {
-		AWSCredentials credentials = new BasicAWSCredentials(accessKey, secretKey);
-		ClientConfiguration clientConfig = new ClientConfiguration();
-		clientConfig.setProtocol(Protocol.HTTP);
-		this.conn = new AmazonS3Client(credentials, clientConfig);
-		conn.setEndpoint("s3.ap-northeast-2.amazonaws.com"); // 엔드포인트 설정 [ 아시아 태평양 서울 ]
-	}
-
-	// 버킷 리스트를 가져오는 메서드이다.
-	public List<Bucket> getBucketList() {
-		return conn.listBuckets();
-	}
-
-	// 버킷을 생성하는 메서드이다.
-	public Bucket createBucket(String bucketName) {
-		return conn.createBucket(bucketName);
-	}
-
-	// 폴더 생성 (폴더는 파일명 뒤에 "/"를 붙여야한다.)
-	public void createFolder(String bucketName, String folderName) {
-		conn.putObject(bucketName, folderName + "/", new ByteArrayInputStream(new byte[0]), new ObjectMetadata());
-	}
-
-	// 파일 업로드
-	public void fileUpload(String bucketName, String fileName, byte[] fileData) throws FileNotFoundException {
-
-		String filePath = (fileName).replace(File.separatorChar, '/'); // 파일 구별자를 `/`로 설정(\->/) 이게 기존에 / 였어도 넘어오면서 \로 바뀌는 거같다.
-		ObjectMetadata metaData = new ObjectMetadata();
-
-		metaData.setContentLength(fileData.length);   //메타데이터 설정 -->원래는 128kB까지 업로드 가능했으나 파일크기만큼 버퍼를 설정시켰다.
-	    ByteArrayInputStream byteArrayInputStream = new ByteArrayInputStream(fileData); //파일 넣음
-
-		conn.putObject(bucketName, filePath, byteArrayInputStream, metaData);
-		
-	}
-
-	// 파일 삭제
-	public void fileDelete(String bucketName, String fileName) {
-		String imgName = (fileName).replace(File.separatorChar, '/');
-		conn.deleteObject(bucketName, imgName);
-		System.out.println("삭제성공");
-	}
-
-	// 파일 URL
-	public String getFileURL(String bucketName, String fileName) {
-		System.out.println("넘어오는 파일명 : "+fileName);
-		String imgName = (fileName).replace(File.separatorChar, '/');
-		return conn.generatePresignedUrl(new GeneratePresignedUrlRequest(bucketName, imgName)).toString();
-	}
-
-}
-
-package org.my.s3;
-
-import java.io.File;
-import java.text.DecimalFormat;
-import java.util.Calendar;
-import java.util.UUID;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class UploadFileUtils {
-
-private static final Logger logger = LoggerFactory.getLogger(UploadFileUtils.class);
-
-
-public static String uploadFile(String uploadPath, String originalName, byte[] byteData) throws Exception {
-	S3Util s3 = new S3Util();
-	String bucketName = "almombucket";
-	//랜덤의 uid 를 만들어준다.
-	UUID uid = UUID.randomUUID();
-
-	//savedName : 570d570a-7af1-4afe-8ed5-391d660084b7_g.JPG 같은 형식으로 만들어준다.
-	String savedName = "/"+uid.toString() + "_" + originalName;
-
-	logger.info("업로드 경로 : "+uploadPath);
-	//\2017\12\27 같은 형태로 저장해준다.
-	String savedPath = calcPath(uploadPath);
-
-	String uploadedFileName = null;
-
-	uploadedFileName = (savedPath + savedName).replace(File.separatorChar, '/');
-	//S3Util 의 fileUpload 메서드로 파일을 업로드한다.
-	s3.fileUpload(bucketName, uploadPath+uploadedFileName, byteData);
-
-
-	logger.info(uploadedFileName);
-//s3.fileUpload(bucketName, new File(fileName))
-
-	return uploadedFileName;
-
-}
-
-private static String calcPath(String uploadPath) {
-
-	Calendar cal = Calendar.getInstance();
-
-	String yearPath = File.separator + cal.get(Calendar.YEAR);
-
-	String monthPath = yearPath + File.separator + new DecimalFormat("00").format(cal.get(Calendar.MONTH) + 1);
-
-	String datePath = monthPath + File.separator + new DecimalFormat("00").format(cal.get(Calendar.DATE));
-
-	makeDir(uploadPath, yearPath, monthPath, datePath);
-
-	logger.info(datePath);
-
-	return datePath;
-}
-
-private static void makeDir(String uploadPath, String... paths) {
-
-	if (new File(paths[paths.length - 1]).exists()) {
-		return;
-	}
-
-	for (String path : paths) {
-
-		File dirPath = new File(uploadPath + path);
-
-		if (!dirPath.exists()) {
-			dirPath.mkdir();
-		}
-	}
-}
-}*/
