+++ dokky/src/main/java/org/my/utils/DownloadView.java
@@ -1,23 +1,12 @@
 package org.my.utils;
-	import java.io.IOException;
-	import java.io.OutputStream;
-	import java.net.URLEncoder;
 	import java.util.Map;
 	import javax.servlet.http.HttpServletRequest;
 	import javax.servlet.http.HttpServletResponse;
 	import org.my.s3.myS3Util;
-import org.my.service.CommonService;
-import org.springframework.beans.factory.annotation.Autowired;
+	import org.my.service.CommonService;
+	import org.springframework.beans.factory.annotation.Autowired;
 	import org.springframework.stereotype.Component;
 	import org.springframework.web.servlet.view.AbstractView;
-	import com.amazonaws.auth.AWSCredentials;
-	import com.amazonaws.auth.AWSStaticCredentialsProvider;
-	import com.amazonaws.auth.BasicAWSCredentials;
-	import com.amazonaws.regions.Regions;
-	import com.amazonaws.services.s3.AmazonS3;
-	import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-	import com.amazonaws.services.s3.model.S3Object;
-	import com.amazonaws.services.s3.model.S3ObjectInputStream;
 	import lombok.Setter;
 		
 @Component
